feat(mobile): 메모 탭 상단에 날씨 배지 및 알림 벨 헤더 추가 (#534)
* feat(mobile): 메모 탭 상단에 날씨 배지 및 알림 벨 헤더 추가 (#533)

* refactor(mobile): WeatherForecastBadge useQuery → useSuspenseQuery 전환 (#533)

* fix(mobile): WeatherForecastBadge useSuspenseQuery → useQuery 복원 (#533)

useSuspenseQuery는 위치 미등록(WEATHER_1902) 같은 예상된 에러도
throw하여 ErrorBoundary 크래시를 유발. useQuery로 복원하여
컴포넌트 내부에서 graceful하게 처리.
